+++ CHANGELOG.md
@@ -101,6 +101,7 @@ New features:
 - Configs can now inherit from other configs through the `parent_config` parameter.
 - Use the `--graph <path>` CLI option to output a graph of the Steps to be run and the dependencies
   between them.
+- Add mode shares by tour count to `AggregateOutputFile`.
 
 Other changes: